+++ docs/conf.py
@@ -24,7 +24,7 @@ copyright = "2023, RocketPy Team"
 author = "RocketPy Team"
 
 # The full version, including alpha/beta/rc tags
-release = "1.1.2"
+release = "1.1.4"
 
 
 # -- General configuration ---------------------------------------------------